Extracted the ProMScrollablePanel from the ProMProperties class so that it can be used in the PromScrollContainer as well.
The ProMScrollContainer is a class in which childs can be added on the fly. These childs have a min/max button so that they can be hidden or displayed when needed.

Further, I added a ProMWizard, which helps in creating wizards for setting configurations. It is a generalization of the TransitionMining wizard.