ETM: Fitness replaying Reusing Alignments
****Fixed bugs related:
-

****Implementation related:
- Major change when repairing the alignments: not considering a block anymore.
- Now the lowest loop dominates the point of change, therefore guidedTreeMutationCoordinator and TreeCrossover select the lowest loop, if present. This can be easily changed: read things to consider (a).
- In treeUtils, added a function to select the lowest parent equal to a given operator.
- In TreeUtils, updated the iterative merge , hence the recursive approach is OUTDATED (never liked the recursive approach anyway). Now the shiftEvent is not a shift, but the actual value.
- In treeUtils, added a function to split a alignment based on the number of log/synch moves. If inputValue>0, it returns the alignment until the sync/log moves are equal to inputValue; If 0, then it returns the alignment until finding a sync/log move
- The 6 steps previously defined to repair an alignment are still the same, but instead of blocks, the algorithm keeps creating a trace until completely exiting the scope.
- When calculating the scope, the algorithm keeps track of the number of consecutive log+sync moves. These are the later index used to split the subalignmantes.
- Added a debug function to "check" the quality of the alignment: check if the log moves are correctly placed.

****Things to test/reconsider
(a) - TODO : It is possible to remove the loop constraint by checking a fresh move inside  the scope, therefore it would be easy to detect the  loops.
(b) -TODO: Precision
(c) -TODO/FIX: RemoveUseless can remove an entire subtree, and the parent can be lost (p.e. XOR()) so if this occurs, it returns the root. 
(d) -TODO: maybe the confidence_level should be more fuzzy, not a simple number. Does the quality is better? worse? Think of this.
(e) -TODO/FIX: What to do when removing useless behavior, all quality criteria should be the same: remove the affected move in the alignment should be enough
(f) -TODO/FIX: Related to the AStarAlgorithm: maybe it will be interesting to create a method that only changes the traces (it requires to remove the final properties of some attributes)

****Things to change/remember:
- Make the HISTORY_CACHE_SIZE dependent on the population size (size = 4*popSize)
- Remember to enable multithreading in AbstractEvolutionEnfine (right now singleThread is forced to true: test and print issues)

****Other minor changes related to the current commit:
- Some changes related to format (Ctrl+shift+F)
- Added more examples in the test plugin.