Bug fix (redo): import log 'logWithProblems.xes'. apply 'mine transition system'. choose all forward and backwards keys. choose 'list' and 'no limit'. choose all options in the following screens. Bug: ProM freezes and doesn't respond to mouse clicks or closing commands anymore. It seems like the computations are done on the gui thread.
This is mainly caused by the size of the resulting transition system. This is simply big (6500 states, 5700 transitions), and requires time to visualize. Nevertheless, the post-mining conversions are also inadequate to deal with such big transition systems. The size of the transition system is now limited to about 200 states. Additional states will be mapped onto one single state, and a message will be shown that the transition system will be truncated. This prevents the user from generating very large transition systems that may cause problems.