Fixed a deadlock problem if the users presses the import button twice fast after each other. As a result, both block each other: The second is blocked because it wants to runs the same synchronized method as the first, the first is blocked as it needs the EDT which is blocked as it is running the second.
The main importResource are now not synchronized anymore, as they rely on the ...NotInEDT methods anyway (which are still synchronized, but do not run in the EDT). Furthermore, these methods now always return true: THe import itself is done in a separate thread, and we cannot wait for the results.